Repacking/renaming.
Moves everything into a new package, `scala.runtime.jfunc`. This is a
preemptive move to avoid split-packaging problems under OSGi.

Uses slightly longer names for the types:

    s/F1/JFunction1/
    s/P1/JProcedure1/
    s/F/JFunc/
