+++ src/test/java/scala/runtime/jfunc/Test.java
@@ -1,11 +1,11 @@
 /*
  * Copyright (C) 2012-2014 Typesafe Inc. <http://www.typesafe.com>
  */
-package scala.runtime.test;
+package scala.runtime.jfunc;
 
 import scala.runtime.*;
-import static scala.runtime.test.TestAPI.*;
-import static scala.runtime.F.*;
+import static scala.runtime.jfunc.TestAPI.*;
+import static scala.runtime.jfunc.JFunc.*;
 
 public class Test {
     public static void main(String[] args) {
@@ -21,11 +21,11 @@ public class Test {
 
         // That's a pity, but we can get pretty close with this library!
 
-        // We have to tell javac to use `F1` as the functional interface.
-        F1<String, String> f1 = (String s) -> s;
+        // We have to tell javac to use `JFunction1` as the functional interface.
+        JFunction1<String, String> f1 = (String s) -> s;
 
         // That's more or less equivalent to the old, anonymous class syntax:
-        new F1<String, String>() {
+        new JFunction1<String, String>() {
             public String apply(String s) { return s; }
         };
 
@@ -40,9 +40,9 @@ public class Test {
         // F1 is a subclass of Function1:
         scala.Function1<String, String> f2 = f1;
 
-        // Factory methods in `F` can reduce the verbosity a little:
+        // Factory methods in `JFunc` can reduce the verbosity a little:
         // `func` is actually just an identity method; it only exists to
-        // trigger lambda creation using the `F1` functional interface.
+        // trigger lambda creation using the `JFunction1` functional interface.
         scala.Function1<String, String> f3 = func((String s) -> s);
 
         // Note that javac's type inference can infer the parameter type here,
@@ -53,7 +53,7 @@ public class Test {
 
         // Specialized variants of the `apply` method are implenented in the
         // functional interface
-        F1<Integer, Integer> f5 = (i) -> -i;
+        JFunction1<Integer, Integer> f5 = (i) -> -i;
         assert(f5.apply(1) == -1);
         assert(f5.apply$mcII$sp(1) == -1);
 
@@ -62,10 +62,10 @@ public class Test {
         scala.Function2<String, String, String> f6 = func((s1, s2) -> join(s1, s2));
         assert(f6.curried().apply("1").apply("2").equals("12"));
 
-        // Functions returning unit must use the `P1`, ... functional interfaces
+        // Functions returning unit must use the `JProcedure1`, ... functional interfaces
         // in order to convert a void lamdba return to Scala's Unit.
         //
-        // The easiest way to do this is via `F.proc`, ....
+        // The easiest way to do this is via `JFunc.proc`, ....
         //
         // Note that the lambda has a return type of `void` if the last
         // statement is a call to a `void` returning method, or if it is
@@ -74,7 +74,7 @@ public class Test {
         scala.Function1<String, BoxedUnit> f8 = proc(s -> {s.toUpperCase(); return;});
 
         // Function0 is also available
-        scala.Function0<String> f9 = F.f0(() -> "42");
+        scala.Function0<String> f9 = func(() -> "42");
         assert(f9.apply().equals("42"));
 
         // You can go up to 22 (the highest arity function defined in the Scala standard library.)
